+++ src/Texture.cpp
@@ -278,7 +278,6 @@ void Texture::beginUse(RenderContext* renderContext)
 
   if (type == ALPHA) {
     glEnable(GL_BLEND);
-    glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA );
   }
 #endif
 }